changed layout of dice page and imported grid button dependency
