+++ rng_flutter/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:rng_flutter/dice.dart';
 
 void main() {
   runApp(const MyApp());
@@ -69,26 +70,26 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 }
 
-class DiceRoute extends StatelessWidget {
-  const DiceRoute({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text("Dice Roller"),
-      ),
-      body: Center(
-        child: ElevatedButton(
-          onPressed: () {
-            Navigator.pop(context);
-          },
-          child: const Text('Go back!'),
-        ),
-      ),
-    );
-  }
-}
+// class DiceRoute extends StatelessWidget {
+//   const DiceRoute({Key? key}) : super(key: key);
+//
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//       appBar: AppBar(
+//         title: const Text("Dice Roller"),
+//       ),
+//       body: Center(
+//         child: ElevatedButton(
+//           onPressed: () {
+//             Navigator.pop(context);
+//           },
+//           child: const Text('Go back!'),
+//         ),
+//       ),
+//     );
+//   }
+// }
 
 class ListRoute extends StatelessWidget {
   const ListRoute({Key? key}) : super(key: key);
