+++ rng_flutter/pubspec.yaml
@@ -29,6 +29,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  flutter_grid_button:
 
 
   # The following adds the Cupertino Icons font to your application.
